[Patch AI Studio v4.9.104] extract numeric helper
